+++ tests/methods/types/methodBool.Test.js
@@ -3,7 +3,7 @@ import { boolData as data } from '../../TestUtil';
 import { testMethodType } from '../methodTestUtility';
 
 describe('method', () => {
-	describe('.bool', () => {
+	describe('.boolean', () => {
 		testMethodType(assign({}, data, {
 			init: false
 		}));