+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
-and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
+
+## Unreleased
+- Changed isBool to isBoolean, enforce.bool to enforce.boolean, and method.bool to method.Boolean
 
 ## 0.2.0 - 2019-01-18
 ### Added